Extract commom behavior from Formatters to a Base class
