+++ lib/awesome_print/formatters/array.rb
@@ -1,18 +1,11 @@
 module AwesomePrint
   module Formatters
-    class Array
-
-      def initialize(formatter, array)
-        @formatter = formatter
-        @array = array
-        @options = formatter.options
-        @inspector = formatter.inspector
-      end
+    class Array < Base
 
       def call
-        return empty_format if array.empty?
+        return empty_format if object.empty?
 
-        if array.instance_variable_defined?('@__awesome_methods__')
+        if object.instance_variable_defined?('@__awesome_methods__')
           methods_format
         elsif options[:multiline]
           multiline_format
@@ -23,32 +16,22 @@ module AwesomePrint
 
       private
 
-        attr_reader :formatter, :array
-
-        def options
-          @options
-        end
-
-        def inspector
-          @inspector
-        end
-
         def empty_format
           '[]'
         end
 
         def methods_format
-          formatter.methods_array(array)
+          formatter.methods_array(object)
         end
 
         def inline_format
-          "[ " << array.map{ |item| inspector.awesome(item) }.join(", ") << " ]"
+          "[ " << object.map{ |item| inspector.awesome(item) }.join(", ") << " ]"
         end
 
         def multiline_format
-          width = (array.size - 1).to_s.size
+          width = (object.size - 1).to_s.size
 
-          data = array.inject([]) do |arr, item|
+          data = object.inject([]) do |arr, item|
             index = formatter.indent
             index << formatter.colorize("[#{arr.size.to_s.rjust(width)}] ", :array) if options[:index]
             formatter.indented do
